refactor: Made changes requested in the PR, moved some functions to the helper file
