+++ front/src/App/components/atom/InputTime.jsx
@@ -2,16 +2,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Box, MaskedInput, Text } from 'grommet';
+import { splitTime } from '#helpers/time';
 
 const pad = (n) => (n < 10 ? '0' : '') + n;
-
-export const splitTime = (time) =>
-  time === undefined ? [0, -1] : time.split(':').map((t) => Number(t));
-
-export const validateTime = (time) => {
-  const timeRegex = /^(0[0-9]|1[0-9]|2[0-3]|[0-9]):[0-5][0-9]$/;
-  return !timeRegex.test(time) && 'Hora inválida';
-};
 
 const InputTime = ({ onChange, start, title, value }) => {
   const [startHour, startMinutes] = splitTime(start);
